[Front | API] Add Confidence in Form or Filter when missing(#issue/2746)
