--- opencti-platform/opencti-front/src/private/components/analysis/opinions/OpinionCreation.js
@@ -1,386 +1,0 @@
-import React, { Component } from 'react';
-import * as PropTypes from 'prop-types';
-import { Field, Form, Formik } from 'formik';
-import { ConnectionHandler } from 'relay-runtime';
-import { compose, evolve, path, pluck } from 'ramda';
-import * as Yup from 'yup';
-import { graphql } from 'react-relay';
-import withStyles from '@mui/styles/withStyles';
-import Drawer from '@mui/material/Drawer';
-import Dialog from '@mui/material/Dialog';
-import DialogContent from '@mui/material/DialogContent';
-import DialogTitle from '@mui/material/DialogTitle';
-import DialogActions from '@mui/material/DialogActions';
-import Typography from '@mui/material/Typography';
-import Button from '@mui/material/Button';
-import IconButton from '@mui/material/IconButton';
-import Fab from '@mui/material/Fab';
-import { Add, Close } from '@mui/icons-material';
-import { commitMutation } from '../../../../relay/environment';
-import inject18n from '../../../../components/i18n';
-import TextField from '../../../../components/TextField';
-import ObjectMarkingField from '../../common/form/ObjectMarkingField';
-import CreatedByField from '../../common/form/CreatedByField';
-import ObjectLabelField from '../../common/form/ObjectLabelField';
-import MarkDownField from '../../../../components/MarkDownField';
-import OpenVocabField from '../../common/form/OpenVocabField';
-import { fieldSpacingContainerStyle } from '../../../../utils/field';
-
-const styles = (theme) => ({
-  drawerPaper: {
-    minHeight: '100vh',
-    width: '50%',
-    position: 'fixed',
-    transition: theme.transitions.create('width', {
-      easing: theme.transitions.easing.sharp,
-      duration: theme.transitions.duration.enteringScreen,
-    }),
-    padding: 0,
-  },
-  createButton: {
-    position: 'fixed',
-    bottom: 30,
-    right: 30,
-  },
-  createButtonContextual: {
-    position: 'fixed',
-    bottom: 30,
-    right: 30,
-    zIndex: 2000,
-  },
-  buttons: {
-    marginTop: 20,
-    textAlign: 'right',
-  },
-  button: {
-    marginLeft: theme.spacing(2),
-  },
-  header: {
-    backgroundColor: theme.palette.background.nav,
-    padding: '20px 20px 20px 60px',
-  },
-  closeButton: {
-    position: 'absolute',
-    top: 12,
-    left: 5,
-    color: 'inherit',
-  },
-  importButton: {
-    position: 'absolute',
-    top: 15,
-    right: 20,
-  },
-  container: {
-    padding: '10px 20px 20px 20px',
-  },
-});
-
-export const opinionCreationMutation = graphql`
-  mutation OpinionCreationMutation($input: OpinionAddInput!) {
-    opinionAdd(input: $input) {
-      ...OpinionLine_node
-    }
-  }
-`;
-
-const opinionValidation = (t) => Yup.object().shape({
-  opinion: Yup.string().required(t('This field is required')),
-  explanation: Yup.string().required(t('This field is required')),
-});
-
-const sharedUpdater = (store, userId, paginationOptions, newEdge) => {
-  const userProxy = store.get(userId);
-  const conn = ConnectionHandler.getConnection(
-    userProxy,
-    'Pagination_opinions',
-    paginationOptions,
-  );
-  ConnectionHandler.insertEdgeBefore(conn, newEdge);
-};
-
-class OpinionCreation extends Component {
-  constructor(props) {
-    super(props);
-    this.state = { open: false };
-  }
-
-  handleOpen() {
-    this.setState({ open: true });
-  }
-
-  handleClose() {
-    this.setState({ open: false });
-  }
-
-  onSubmit(values, { setSubmitting, resetForm }) {
-    const adaptedValues = evolve(
-      {
-        createdBy: path(['value']),
-        objectMarking: pluck('value'),
-        objectLabel: pluck('value'),
-      },
-      values,
-    );
-    commitMutation({
-      mutation: opinionCreationMutation,
-      variables: {
-        input: adaptedValues,
-      },
-      updater: (store) => {
-        const payload = store.getRootField('opinionAdd');
-        const newEdge = payload.setLinkedRecord(payload, 'node'); // Creation of the pagination container.
-        const container = store.getRoot();
-        sharedUpdater(
-          store,
-          container.getDataID(),
-          this.props.paginationOptions,
-          newEdge,
-        );
-      },
-      setSubmitting,
-      onCompleted: () => {
-        setSubmitting(false);
-        resetForm();
-        this.handleClose();
-      },
-    });
-  }
-
-  onResetClassic() {
-    this.handleClose();
-  }
-
-  onResetContextual() {
-    this.handleClose();
-  }
-
-  renderClassic() {
-    const { t, classes } = this.props;
-    return (
-      <div>
-        <Fab
-          onClick={this.handleOpen.bind(this)}
-          color="secondary"
-          aria-label="Add"
-          className={classes.createButton}
-        >
-          <Add />
-        </Fab>
-        <Drawer
-          open={this.state.open}
-          anchor="right"
-          elevation={1}
-          sx={{ zIndex: 1202 }}
-          classes={{ paper: classes.drawerPaper }}
-          onClose={this.handleClose.bind(this)}
-        >
-          <div className={classes.header}>
-            <IconButton
-              aria-label="Close"
-              className={classes.closeButton}
-              onClick={this.handleClose.bind(this)}
-              size="large"
-              color="primary"
-            >
-              <Close fontSize="small" color="primary" />
-            </IconButton>
-            <Typography variant="h6">{t('Create a opinions')}</Typography>
-          </div>
-          <div className={classes.container}>
-            <Formik
-              initialValues={{
-                opinion: '',
-                explanation: '',
-                createdBy: '',
-                objectMarking: [],
-                objectLabel: [],
-              }}
-              validationSchema={opinionValidation(t)}
-              onSubmit={this.onSubmit.bind(this)}
-              onReset={this.onResetClassic.bind(this)}
-            >
-              {({
-                submitForm,
-                handleReset,
-                isSubmitting,
-                setFieldValue,
-                values,
-              }) => (
-                <Form style={{ margin: '20px 0 20px 0' }}>
-                  <OpenVocabField
-                    label={t('Opinion')}
-                    type="opinion_ov"
-                    name="opinion"
-                    onChange={(name, value) => setFieldValue(name, value)}
-                    containerStyle={fieldSpacingContainerStyle}
-                    multiple={false}
-                  />
-                  <Field
-                    component={TextField}
-                    variant="standard"
-                    name="explanation"
-                    label={t('Explanation')}
-                    fullWidth={true}
-                    multiline={true}
-                    rows="4"
-                    style={{ marginTop: 20 }}
-                  />
-                  <CreatedByField
-                    name="createdBy"
-                    style={{ marginTop: 20, width: '100%' }}
-                    setFieldValue={setFieldValue}
-                  />
-                  <ObjectLabelField
-                    name="objectLabel"
-                    style={{ marginTop: 20, width: '100%' }}
-                    setFieldValue={setFieldValue}
-                    values={values.objectLabel}
-                  />
-                  <ObjectMarkingField
-                    name="objectMarking"
-                    style={{ marginTop: 20, width: '100%' }}
-                  />
-                  <div className={classes.buttons}>
-                    <Button
-                      variant="contained"
-                      onClick={handleReset}
-                      disabled={isSubmitting}
-                      classes={{ root: classes.button }}
-                    >
-                      {t('Cancel')}
-                    </Button>
-                    <Button
-                      variant="contained"
-                      color="secondary"
-                      onClick={submitForm}
-                      disabled={isSubmitting}
-                      classes={{ root: classes.button }}
-                    >
-                      {t('Create')}
-                    </Button>
-                  </div>
-                </Form>
-              )}
-            </Formik>
-          </div>
-        </Drawer>
-      </div>
-    );
-  }
-
-  renderContextual() {
-    const { t, classes, inputValue, display } = this.props;
-    return (
-      <div style={{ display: display ? 'block' : 'none' }}>
-        <Fab
-          onClick={this.handleOpen.bind(this)}
-          color="secondary"
-          aria-label="Add"
-          className={classes.createButtonContextual}
-        >
-          <Add />
-        </Fab>
-        <Dialog
-          open={this.state.open}
-          onClose={this.handleClose.bind(this)}
-          PaperProps={{ elevation: 1 }}
-        >
-          <Formik
-            enableReinitialize={true}
-            initialValues={{
-              opinion: inputValue,
-              explanation: '',
-              createdBy: '',
-              objectMarking: [],
-              objectLabel: [],
-            }}
-            validationSchema={opinionValidation(t)}
-            onSubmit={this.onSubmit.bind(this)}
-            onReset={this.onResetContextual.bind(this)}
-          >
-            {({
-              submitForm,
-              handleReset,
-              isSubmitting,
-              setFieldValue,
-              values,
-            }) => (
-              <Form>
-                <DialogTitle>{t('Create an opinion')}</DialogTitle>
-                <DialogContent>
-                  <OpenVocabField
-                    label={t('Opinion')}
-                    type="opinion_ov"
-                    name="opinion"
-                    onChange={(name, value) => setFieldValue(name, value)}
-                    containerStyle={fieldSpacingContainerStyle}
-                    multiple={false}
-                  />
-                  <Field
-                    component={MarkDownField}
-                    name="explanation"
-                    label={t('Explanation')}
-                    fullWidth={true}
-                    multiline={true}
-                    rows="4"
-                    style={{ marginTop: 20 }}
-                  />
-                  <CreatedByField
-                    name="createdBy"
-                    style={fieldSpacingContainerStyle}
-                    setFieldValue={setFieldValue}
-                  />
-                  <ObjectLabelField
-                    name="objectLabel"
-                    style={fieldSpacingContainerStyle}
-                    setFieldValue={setFieldValue}
-                    values={values.objectLabel}
-                  />
-                  <ObjectMarkingField
-                    name="objectMarking"
-                    style={fieldSpacingContainerStyle}
-                  />
-                </DialogContent>
-                <DialogActions>
-                  <Button onClick={handleReset} disabled={isSubmitting}>
-                    {t('Cancel')}
-                  </Button>
-                  <Button
-                    color="secondary"
-                    onClick={submitForm}
-                    disabled={isSubmitting}
-                  >
-                    {t('Create')}
-                  </Button>
-                </DialogActions>
-              </Form>
-            )}
-          </Formik>
-        </Dialog>
-      </div>
-    );
-  }
-
-  render() {
-    const { contextual } = this.props;
-    if (contextual) {
-      return this.renderContextual();
-    }
-    return this.renderClassic();
-  }
-}
-
-OpinionCreation.propTypes = {
-  paginationOptions: PropTypes.object,
-  classes: PropTypes.object,
-  theme: PropTypes.object,
-  t: PropTypes.func,
-  contextual: PropTypes.bool,
-  display: PropTypes.bool,
-  inputValue: PropTypes.string,
-};
-
-export default compose(
-  inject18n,
-  withStyles(styles, { withTheme: true }),
-)(OpinionCreation);
