+++ opencti-platform/opencti-front/src/private/components/analysis/opinions/StixCoreObjectOpinionsRadar.js
@@ -26,6 +26,8 @@ import MarkDownField from '../../../../components/MarkDownField';
 import { adaptFieldValue } from '../../../../utils/String';
 import { opinionMutationFieldPatch } from './OpinionEditionOverview';
 import { radarChartOptions } from '../../../../utils/Charts';
+import { fieldSpacingContainerStyle } from '../../../../utils/field';
+import ConfidenceField from '../../common/form/ConfidenceField';
 
 const styles = () => ({
   paper: {
@@ -106,6 +108,7 @@ class StixCoreObjectOpinionsRadarComponent extends Component {
     if (alreadyExistingOpinion) {
       const inputValues = R.pipe(
         R.dissoc('alreadyExistingOpinion'),
+        R.assoc('confidence', parseInt(values.confidence, 10)),
         R.assoc('opinion', opinions[(currentOpinion || 3) - 1]),
         R.assoc('objectMarking', defaultMarking),
         R.toPairs,
@@ -129,6 +132,7 @@ class StixCoreObjectOpinionsRadarComponent extends Component {
     } else {
       const adaptedValues = R.pipe(
         R.dissoc('alreadyExistingOpinion'),
+        R.assoc('confidence', parseInt(values.confidence, 10)),
         R.assoc('opinion', opinions[(currentOpinion || 3) - 1]),
         R.assoc('objectMarking', defaultMarking),
         R.assoc('objects', [stixCoreObjectId]),
@@ -299,6 +303,7 @@ class StixCoreObjectOpinionsRadarComponent extends Component {
                       initialValues={{
                         alreadyExistingOpinion: props.myOpinion?.id || '',
                         explanation,
+                        confidence: 75,
                       }}
                       onSubmit={this.onSubmit.bind(this)}
                       onReset={this.onReset.bind(this)}
@@ -327,6 +332,12 @@ class StixCoreObjectOpinionsRadarComponent extends Component {
                               multiline={true}
                               rows="4"
                               style={{ marginTop: 20 }}
+                            />
+                            <ConfidenceField
+                              name="confidence"
+                              label={t('Confidence')}
+                              fullWidth={true}
+                              containerStyle={fieldSpacingContainerStyle}
                             />
                           </DialogContent>
                           <DialogActions>
