+++ opencti-platform/opencti-front/src/private/components/arsenal/Malwares.js
@@ -147,6 +147,7 @@ class Malwares extends Component {
           'createdBy',
           'revoked',
           'malware_types',
+          'confidence',
         ]}
       >
         <QueryRenderer
@@ -220,6 +221,7 @@ class Malwares extends Component {
           'created_start_date',
           'created_end_date',
           'createdBy',
+          'confidence',
         ]}
       >
         <QueryRenderer
