+++ opencti-platform/opencti-front/src/private/components/arsenal/Tools.js
@@ -148,6 +148,7 @@ class Tools extends Component {
           'created_start_date',
           'created_end_date',
           'createdBy',
+          'confidence',
         ]}
       >
         <QueryRenderer
