+++ opencti-platform/opencti-front/src/private/components/arsenal/channels/ChannelLine.js
@@ -112,6 +112,7 @@ export const ChannelLine = createFragmentContainer(ChannelLineComponent, {
       channel_types
       created
       modified
+      confidence
       objectLabel {
         edges {
           node {
