+++ opencti-platform/opencti-front/src/private/components/arsenal/malwares/MalwareLine.js
@@ -115,6 +115,7 @@ const MalwareLineFragment = createFragmentContainer(MalwareLineComponent, {
       name
       created
       modified
+      confidence
       objectMarking {
         edges {
           node {
