+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilityCreation.tsx
@@ -21,6 +21,8 @@ import { Theme } from '../../../../components/Theme';
 import { VulnerabilitiesLinesPaginationQuery$variables } from './__generated__/VulnerabilitiesLinesPaginationQuery.graphql';
 import { Option } from '../../common/form/ReferenceField';
 import { insertNode } from '../../../../utils/store';
+import { fieldSpacingContainerStyle } from '../../../../utils/field';
+import ConfidenceField from '../../common/form/ConfidenceField';
 
 const useStyles = makeStyles<Theme>((theme) => ({
   drawerPaper: {
@@ -82,6 +84,7 @@ const vulnerabilityValidation = (t: (message: string) => string) => Yup.object()
   x_opencti_base_score: Yup.number(),
   x_opencti_base_severity: Yup.string().nullable(),
   x_opencti_attack_vector: Yup.string().nullable(),
+  confidence: Yup.number(),
 });
 
 interface VulnerabilityAddInput {
@@ -93,6 +96,7 @@ interface VulnerabilityAddInput {
   x_opencti_base_score: number;
   x_opencti_base_severity: string;
   x_opencti_attack_vector: string;
+  confidence: number;
 }
 
 interface VulnerabilityCreationProps {
@@ -121,6 +125,7 @@ const VulnerabilityCreation: FunctionComponent<VulnerabilityCreationProps> = ({
     x_opencti_base_score: 1,
     x_opencti_base_severity: '',
     x_opencti_attack_vector: '',
+    confidence: 75,
   };
 
   const [commit] = useMutation(vulnerabilityMutation);
@@ -142,6 +147,7 @@ const VulnerabilityCreation: FunctionComponent<VulnerabilityCreationProps> = ({
       x_opencti_base_score: values.x_opencti_base_score,
       x_opencti_base_severity: values.x_opencti_base_severity,
       x_opencti_attack_vector: values.x_opencti_attack_vector,
+      confidence: parseInt(String(values.confidence), 10),
     };
     commit({
       variables: {
@@ -216,6 +222,12 @@ const VulnerabilityCreation: FunctionComponent<VulnerabilityCreationProps> = ({
         label={t('CVSS3 - Attack vector')}
         fullWidth={true}
         style={{ marginTop: 20 }}
+      />
+      <ConfidenceField
+        name="confidence"
+        label={t('Confidence')}
+        fullWidth={true}
+        containerStyle={fieldSpacingContainerStyle}
       />
       <CreatedByField
         name="createdBy"
