+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilityEditionOverview.js
@@ -121,6 +121,7 @@ const vulnerabilityValidation = (t) => Yup.object().shape({
     .required(t('This field is required')),
   references: Yup.array().required(t('This field is required')),
   x_opencti_workflow_id: Yup.object(),
+  confidence: Yup.number(),
 });
 
 class VulnerabilityEditionOverviewComponent extends Component {
@@ -382,6 +383,7 @@ const VulnerabilityEditionOverview = createFragmentContainer(
         id
         name
         description
+        confidence
         createdBy {
           ... on Identity {
             id
