+++ opencti-platform/opencti-front/src/private/components/arsenal/vulnerabilities/VulnerabilityLine.tsx
@@ -63,6 +63,7 @@ const vulnerabilityLineFragment = graphql`
     x_opencti_base_severity
     created
     modified
+    confidence
     objectLabel {
       edges {
         node {
