+++ opencti-platform/opencti-front/src/private/components/events/StixSightingRelationships.tsx
@@ -106,6 +106,7 @@ const StixSightingRelationships = () => {
         'createdBy',
         'toSightingId',
         'x_opencti_negative',
+        'confidence',
       ]}
     >
       <QueryRenderer
