+++ opencti-platform/opencti-front/src/private/components/events/incidents/IncidentLine.js
@@ -168,6 +168,7 @@ const IncidentLineFragment = createFragmentContainer(IncidentLineComponent, {
       severity
       created
       modified
+      confidence
       objectMarking {
         edges {
           node {
