+++ opencti-platform/opencti-front/src/private/components/observations/Indicators.js
@@ -350,6 +350,7 @@ class Indicators extends Component {
                 'basedOn',
                 'revoked',
                 'creator',
+                'confidence',
               ]}
             >
               <QueryRenderer
