+++ opencti-platform/opencti-front/src/private/components/observations/Infrastructures.js
@@ -150,6 +150,7 @@ class Infrastructures extends Component {
           'created_start_date',
           'created_end_date',
           'createdBy',
+          'confidence',
         ]}
       >
         <QueryRenderer
