+++ opencti-platform/opencti-front/src/private/components/observations/infrastructures/InfrastructureLine.js
@@ -118,6 +118,7 @@ const InfrastructureLineFragment = createFragmentContainer(
         name
         created
         modified
+        confidence
         objectMarking {
           edges {
             node {
