+++ opencti-platform/opencti-front/src/private/components/techniques/DataComponents.tsx
@@ -93,6 +93,7 @@ const DataComponents: FunctionComponent = () => {
           'created_end_date',
           'createdBy',
           'revoked',
+          'confidence',
         ]}
       >
         {queryRef && (
