+++ opencti-platform/opencti-front/src/private/components/techniques/DataSources.tsx
@@ -90,6 +90,7 @@ const DataSources: FunctionComponent = () => {
           'created_end_date',
           'createdBy',
           'revoked',
+          'confidence',
         ]}
       >
         {queryRef && (
