+++ opencti-platform/opencti-front/src/private/components/techniques/data_components/DataComponentLine.tsx
@@ -62,6 +62,7 @@ const dataComponentFragment = graphql`
     description
     created
     modified
+    confidence
     objectLabel {
       edges {
         node {
