+++ opencti-platform/opencti-front/src/private/components/techniques/data_sources/DataSourceLine.tsx
@@ -65,6 +65,7 @@ const dataSourceLineFragment = graphql`
     modified
     x_mitre_platforms
     collection_layers
+    confidence
     objectLabel {
       edges {
         node {
