+++ opencti-platform/opencti-front/src/private/components/threats/IntrusionSets.js
@@ -149,6 +149,7 @@ class IntrusionSets extends Component {
           'created_end_date',
           'createdBy',
           'revoked',
+          'confidence',
         ]}
       >
         <QueryRenderer
@@ -222,6 +223,7 @@ class IntrusionSets extends Component {
           'created_start_date',
           'created_end_date',
           'createdBy',
+          'confidence',
         ]}
       >
         <QueryRenderer
