+++ opencti-platform/opencti-front/src/private/components/threats/ThreatActors.js
@@ -149,6 +149,7 @@ class ThreatActors extends Component {
           'created_end_date',
           'createdBy',
           'revoked',
+          'confidence',
         ]}
       >
         <QueryRenderer
@@ -223,6 +224,7 @@ class ThreatActors extends Component {
           'created_start_date',
           'created_end_date',
           'createdBy',
+          'confidence',
         ]}
       >
         <QueryRenderer
