+++ opencti-platform/opencti-front/src/private/components/threats/intrusion_sets/IntrusionSetLine.js
@@ -118,6 +118,7 @@ const IntrusionSetLineFragment = createFragmentContainer(
         name
         created
         modified
+        confidence
         objectMarking {
           edges {
             node {
