+++ opencti-platform/opencti-front/src/private/components/threats/threat_actors/ThreatActorLine.js
@@ -118,6 +118,7 @@ const ThreatActorLineFragment = createFragmentContainer(
         name
         created
         modified
+        confidence
         objectLabel {
           edges {
             node {
