+++ opencti-platform/opencti-front/src/schema/relay.schema.graphql
@@ -2160,6 +2160,7 @@ enum CampaignsOrdering {
   created_at
   updated_at
   x_opencti_workflow_id
+  confidence
 }
 
 enum CampaignsFilter {
@@ -2175,6 +2176,7 @@ enum CampaignsFilter {
   labelledBy
   x_opencti_workflow_id
   revoked
+  confidence
 }
 
 input CampaignsFiltering {
@@ -2510,6 +2512,7 @@ enum ObservedDatasOrdering {
   createdBy
   x_opencti_workflow_id
   objectMarking
+  confidence
 }
 
 enum ObservedDatasFilter {
@@ -2527,6 +2530,7 @@ enum ObservedDatasFilter {
   objectContains
   x_opencti_workflow_id
   revoked
+  confidence
 }
 
 input ObservedDatasFiltering {
@@ -3639,6 +3643,7 @@ enum IndicatorsFilter {
   sightedBy
   revoked
   creator
+  confidence
 }
 
 input IndicatorsFiltering {
@@ -3756,6 +3761,7 @@ enum InfrastructuresOrdering {
   created_at
   updated_at
   x_opencti_workflow_id
+  confidence
 }
 
 enum InfrastructuresFilter {
@@ -3774,6 +3780,7 @@ enum InfrastructuresFilter {
   labelledBy
   x_opencti_workflow_id
   revoked
+  confidence
 }
 
 input InfrastructuresFiltering {
@@ -3874,6 +3881,7 @@ enum IntrusionSetsOrdering {
   created_at
   updated_at
   x_opencti_workflow_id
+  confidence
 }
 
 enum IntrusionSetsFilter {
@@ -3889,6 +3897,7 @@ enum IntrusionSetsFilter {
   labelledBy
   x_opencti_workflow_id
   revoked
+  confidence
 }
 
 input IntrusionSetsFiltering {
@@ -4589,6 +4598,7 @@ enum MalwaresOrdering {
   objectMarking
   objectLabel
   x_opencti_workflow_id
+  confidence
 }
 
 enum MalwaresFilter {
@@ -4607,6 +4617,7 @@ enum MalwaresFilter {
   assigneeTo
   labelledBy
   x_opencti_workflow_id
+  confidence
 }
 
 input MalwaresFiltering {
@@ -4716,6 +4727,7 @@ enum ThreatActorsOrdering {
   created_at
   updated_at
   x_opencti_workflow_id
+  confidence
 }
 
 enum ThreatActorsFilter {
@@ -4731,6 +4743,7 @@ enum ThreatActorsFilter {
   labelledBy
   x_opencti_workflow_id
   revoked
+  confidence
 }
 
 input ThreatActorsFiltering {
@@ -4846,6 +4859,7 @@ enum ToolsOrdering {
   created_at
   updated_at
   x_opencti_workflow_id
+  confidence
 }
 
 enum ToolsFilter {
@@ -4859,6 +4873,7 @@ enum ToolsFilter {
   assigneeTo
   labelledBy
   x_opencti_workflow_id
+  confidence
 }
 
 input ToolsFiltering {
@@ -4960,6 +4975,7 @@ enum VulnerabilitiesOrdering {
   updated_at
   x_opencti_workflow_id
   creator
+  confidence
 }
 
 enum VulnerabilitiesFilter {
@@ -4977,6 +4993,7 @@ enum VulnerabilitiesFilter {
   labelledBy
   x_opencti_workflow_id
   creator
+  confidence
 }
 
 input VulnerabilitiesFiltering {
@@ -5087,6 +5104,7 @@ enum IncidentsOrdering {
   updated_at
   x_opencti_workflow_id
   objectMarking
+  confidence
 }
 
 enum IncidentsFilter {
@@ -5103,6 +5121,7 @@ enum IncidentsFilter {
   assigneeTo
   labelledBy
   x_opencti_workflow_id
+  confidence
 }
 
 input IncidentsFiltering {
@@ -8369,6 +8388,7 @@ enum ChannelsOrdering {
   objectMarking
   objectLabel
   x_opencti_workflow_id
+  confidence
 }
 
 enum ChannelsFilter {
@@ -8384,6 +8404,7 @@ enum ChannelsFilter {
   assigneeTo
   labelledBy
   x_opencti_workflow_id
+  confidence
 }
 
 input ChannelsFiltering {
@@ -8932,6 +8953,7 @@ enum DataComponentsOrdering {
   created_at
   updated_at
   x_opencti_workflow_id
+  confidence
 }
 
 enum DataComponentsFilter {
@@ -8946,6 +8968,7 @@ enum DataComponentsFilter {
   labelledBy
   revoked
   x_opencti_workflow_id
+  confidence
 }
 
 input DataComponentsFiltering {
@@ -9042,6 +9065,7 @@ enum DataSourcesOrdering {
   created_at
   updated_at
   x_opencti_workflow_id
+  confidence
 }
 
 enum DataSourcesFilter {
@@ -9056,6 +9080,7 @@ enum DataSourcesFilter {
   labelledBy
   revoked
   x_opencti_workflow_id
+  confidence
 }
 
 input DataSourcesFiltering {
@@ -9277,6 +9302,7 @@ enum CasesOrdering {
   markedBy
   x_opencti_workflow_id
   case_type
+  confidence
 }
 
 enum CasesFilter {
@@ -9294,6 +9320,7 @@ enum CasesFilter {
   x_opencti_workflow_id
   rating
   case_type
+  confidence
 }
 
 input CasesFiltering {
