+++ opencti-platform/opencti-graphql/src/modules/case/case.ts
@@ -37,6 +37,7 @@ const CASE_DEFINITION: ModuleDefinition<StoreEntityCase> = {
     { name: 'x_opencti_workflow_id', type: 'string', multiple: false, upsert: true },
     { name: 'rating', type: 'numeric', multiple: false, upsert: true },
     { name: 'x_opencti_stix_ids', type: 'string', multiple: true, upsert: true },
+    { name: 'confidence', type: 'numeric', multiple: false, upsert: true },
   ],
   relations: [],
   converter: convertCaseToStix
