+++ opencti-platform/opencti-graphql/src/modules/dataComponent/dataComponent.graphql
@@ -123,6 +123,7 @@ enum DataComponentsOrdering {
     created_at
     updated_at
     x_opencti_workflow_id
+    confidence
 }
 
 # Filtering
@@ -138,6 +139,7 @@ enum DataComponentsFilter {
     labelledBy
     revoked
     x_opencti_workflow_id
+    confidence
 }
 
 input DataComponentsFiltering {
