+++ opencti-platform/opencti-graphql/src/modules/dataSource/dataSource.graphql
@@ -124,6 +124,7 @@ enum DataSourcesOrdering {
     created_at
     updated_at
     x_opencti_workflow_id
+    confidence
 }
 
 # Filtering
@@ -139,6 +140,7 @@ enum DataSourcesFilter {
     labelledBy
     revoked
     x_opencti_workflow_id
+    confidence
 }
 
 input DataSourcesFiltering {
